fix: 당첨 금액 오버플로우 문제 해결

+++ src/main/java/lotto/model/result/Profit.java
@@ -3,9 +3,9 @@ package lotto.model.result;
 public class Profit {
     private static final int INIT = 0;
 
-    private int number;
+    private long number;
 
-    public int sumProfitMoney(WinningResult winningResult) {
+    public long sumProfitMoney(WinningResult winningResult) {
         init();
         for (Rank rank : winningResult.getWinningCount().keySet()) {
             number += winningResult.getWinningCount().get(rank) * rank.getValue();
